// src/api/model/createCompanyDtoCompanyType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export type CreateCompanyDtoCompanyType = typeof CreateCompanyDtoCompanyType[keyof typeof CreateCompanyDtoCompanyType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateCompanyDtoCompanyType = {
  OSVC: 'OSVC',
  SRO: 'SRO',
} as const;